# docs/swagger.yaml
basePath: /
definitions:
  handler.ErrorDetail:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  handler.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/handler.ErrorDetail'
      success:
        example: false
        type: boolean
    type: object
  handler.SuccessResponse-array_models_Category:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Category'
        type: array
      success:
        example: true
        type: boolean
    type: object
...
  handler.SuccessResponse-handler_loginResponse:
    properties:
      data:
        $ref: '#/definitions/handler.loginResponse'
      success:
        example: true
        type: boolean
    type: object
  handler.SuccessResponse-handler_logoutResponse:
    properties:
      data:
        $ref: '#/definitions/handler.logoutResponse'
      success:
        example: true
        type: boolean
    type: object
  handler.SuccessResponse-handler_registerResponse:
    properties:
      data:
        $ref: '#/definitions/handler.registerResponse'
      success:
        example: true
        type: boolean
    type: object
  handler.SuccessResponse-handler_updateCartResponse:
    properties:
      data:
        $ref: '#/definitions/handler.updateCartResponse'
      success:
        example: true
        type: boolean
    type: object
  handler.SuccessResponse-models_Cart:
    properties:
      data:
        $ref: '#/definitions/models.Cart'
      success:
        example: true
        type: boolean
    type: object
  handler.SuccessResponse-models_Category:
    properties:
      data:
        $ref: '#/definitions/models.Category'
      success:
        example: true
        type: boolean
    type: object
...
  handler.addCartItemRequest:
    properties:
      product_id:
        type: integer
      quantity:
        type: integer
    type: object
  handler.createCategoryRequest:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
...
  handler.loginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  handler.loginResponse:
    properties:
      access_token:
        type: string
      expires_in_seconds:
        type: integer
      refresh_token:
        type: string
    type: object
  handler.logoutResponse:
    properties:
      message:
        example: successfully logged out
        type: string
    type: object
...
  handler.refreshTokenRequest:
    properties:
      refresh_token:
        type: string
    type: object
  handler.registerRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      role:
        type: string
    type: object
  handler.registerResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      name:
        type: string
      user_id:
        type: integer
    type: object
  handler.updateCartResponse:
    properties:
      message:
        example: cart item updated successfully
        type: string
    type: object
...
  models.Cart:
    properties:
      created_at:
        type: string
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/models.CartItem'
        type: array
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  models.CartItem:
    properties:
      price:
        type: integer
      product_id:
        type: integer
      product_name:
        type: string
      quantity:
        type: integer
    type: object
  models.Category:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: ShopFlow modern backend engineering showcase API in Go.
  title: ShopFlow API
  version: "1.0"
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Log in with email and password to receive JWT tokens.
      parameters:
      - description: User login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successful login
          schema:
            $ref: '#/definitions/handler.SuccessResponse-handler_loginResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Authenticate user
      tags:
      - Auth
  /api/v1/auth/logout:
    post:
      consumes:
      - application/json
      description: Invalidate the session refresh token.
      parameters:
      - description: Refresh token to invalidate
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.refreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully logged out
          schema:
            $ref: '#/definitions/handler.SuccessResponse-handler_logoutResponse'
        "400":
          description: Invalid refresh token or token missing
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Log out user
      tags:
      - Auth
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: Rotate expired access token using a valid refresh token.
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.refreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Tokens refreshed successfully
          schema:
            $ref: '#/definitions/handler.SuccessResponse-handler_loginResponse'
        "400":
          description: Invalid refresh token or token missing
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Refresh JWT tokens
      tags:
      - Auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Create a customer or admin account.
      parameters:
      - description: User registration details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.registerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            $ref: '#/definitions/handler.SuccessResponse-handler_registerResponse'
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: Email already exists
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Register a new user
      tags:
      - Auth
  /api/v1/cart:
    get:
      description: Retrieve the authenticated user's active shopping cart. Authenticated.
      produces:
      - application/json
      responses:
        "200":
          description: Cart details retrieved successfully
          schema:
            $ref: '#/definitions/handler.SuccessResponse-models_Cart'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: View shopping cart
      tags:
      - Cart
  /api/v1/cart/items:
    post:
      consumes:
      - application/json
      description: Add a product to the user's shopping cart or update its quantity.
        Authenticated.
      parameters:
      - description: Cart item details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.addCartItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Item added/updated successfully
          schema:
            $ref: '#/definitions/handler.SuccessResponse-handler_updateCartResponse'
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: Insufficient stock
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add or update item in cart
      tags:
      - Cart
  /api/v1/cart/items/{productId}:
    delete:
      description: Remove a product from the user's shopping cart completely. Authenticated.
      parameters:
      - description: Product ID
        in: path
        name: productId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid product ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Cart or product not found in cart
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove item from cart
      tags:
      - Cart
  /api/v1/categories:
    get:
      description: Retrieve a list of all product categories. Public.
      produces:
      - application/json
      responses:
        "200":
          description: Categories retrieved successfully
          schema:
            $ref: '#/definitions/handler.SuccessResponse-array_models_Category'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: List categories
      tags:
      - Category
    post:
      consumes:
      - application/json
      description: Create a new product category. Admin only.
      parameters:
      - description: Category details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.createCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Category created successfully
          schema:
            $ref: '#/definitions/handler.SuccessResponse-models_Category'
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: Category already exists
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create category
      tags:
      - Category
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer <your-jwt-token>" to authenticate.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"